requirements.txt
python-dotenv>=1.0.0
customtkinter>=5.2.0
pystray
Pillow
playwright>=1.40.0
plyer
requests
pyautogui
keyboard
opencv-python
numpy
psutil>=5.9.0
wmi>=1.5.1
pywin32>=306
selenium>=4.10.0
webdriver-manager>=4.0.0